Fix some more warnings from dependencies

+++ tests/test_recharge.py
@@ -9,7 +9,8 @@ evap = read_csv("tests/data/evap.csv", index_col=0,
                 parse_dates=True).squeeze("columns").loc["2005":] * 1e3
 obs = read_csv("tests/data/obs.csv", index_col=0,
                parse_dates=True).squeeze("columns")
-temp = Series(index=evap.index, data=sin(arange(evap.size) / 365 * 6))
+temp = Series(index=evap.index, data=sin(arange(evap.size) / 365 * 6),
+              dtype=float)
 
 
 def test_create_rechargemodel():
